PageRank test on star-like graphs.
This commit adds test_pagerank_on_cugraph_digraph_via_star_like_graphs, which tests page rank on graphs with a star-like structure.

The graphs have a hub that all outer nodes point to. A fraction of the outer nodes point to a neighboring outer node.

This test checks that the hub has the highest rank.