Update PageRank tests
This commit removes all of the old PageRank tests, i.e.
* test_pagerank_on_cugraph_digraph_via_tiny_fully_connected_graph
* test_pagerank_on_cugraph_digraph_via_fully_connected_graphs
* test_pagerank_on_cugraph_digraph_via_star_like_graphs

Those tests tested the backend implementations of PageRank rather than metagraph-cuda itself. Testing the implementation of PageRank is out of scope since the implementation is managed via efforts outside of metagraph-cuda.

We instead added the test test_pagerank_on_cugraph_digraph that tests functionality that is within the scope of metagraph-cuda.

This test does the following:
* Verify that PageRank is callable.
* Verify that cugraph.DiGraph is supported directly in the resolver.
* Verify that the PageRank algorithm is present.
* Verify that PageRank returns the desired output type (i.e. numpy.ndarray).