Rework metagraph-cuda to use abstract properties.
This commit makes metagraph-cuda use the abstract properties implemented in https://github.com/ContinuumIO/metagraph/pull/28 as of commit https://github.com/ContinuumIO/metagraph/pull/28/commits/e60839e8c83ad4538fe8c1e5ea28261696daa9fc .

This commit includes a pervasive reuse of the new NumpyNodes in place of NumpySparseVector.

This commit also updates some of the tests to be inline with the changes made to support abstract properties.

This commit updates the triangle count tests to work on undirected graphs only.

This commit includes a rename of CuGraphType -> AutoCuGraphType to avoid name collisions with the new-as-of-this-commit CuGraph wrapper.

This commit ports CuDFEdgeList, AutoCuGraphType (formerly known as CuGraphType), and CuGraph.

This commit introduces AutoCuDiGraphType.

This commit gets rid of CuDFWeightedEdgeList and CuGraphWeightedsince weightedness is now an abstract property.