Remove Auto* Concrete Types
This commit removes AutoCuDiGraphType and AutoCuGraphType since it's been decided that it'll be a requirement for graphs to always be wrapped.

This commit updates the tests to wrap the graphs used in the appropriate wrappers instead of relying on the existence of AutoCuDiGraphType and AutoCuGraphType.

This commit updates the concrete_algorithm cugraph_pagerank to take a graph wrapped in CuGraph instead of a cugraph.DiGraph instance since we removed AutoCuDiGraphType and AutoCuGraphType.

The analogous changes were made to cugraph_triangle_count, which was formerly known as auto_cugraph_triangle_count prior to this commit.

This commit removes translate_graph_cugraph2autocugraph as well since AutoCuGraphType no longer exists.

The same goes for translate_graph_cudfedge2cugraph.

This patch also removes the helper functions _determine_dtype_from_cugraph_graph and _determine_weights_from_cugraph_graph from metagraph_cuda/types.py since those helpers were only used to reduce the redundant code shared between AutoCuDiGraphType, AutoCuGraphType, and CuGraph. Since CuGraph is the only user of those helpers, their internals have been inlined in the definition of CuGraph.